+++ EscapeRoom/EscapeRoom/src/Highscore.cpp
@@ -50,7 +50,7 @@ void Highscore::handleInput(sf::Event& event, sf::RenderWindow& window, Game& ga
     }
     else if (event.type == sf::Event::MouseButtonPressed) {
         if (event.mouseButton.button == sf::Mouse::Left) {
-            sf::Vector2f clickPosition(event.mouseButton.x, event.mouseButton.y);
+            sf::Vector2f clickPosition(static_cast<float>(event.mouseButton.x), static_cast<float>(event.mouseButton.y));
 
             if (translated_pos.x >= 39 && translated_pos.x <= 361 && translated_pos.y >= 525 && translated_pos.y <= 575) {
                 AudioManager::getInstance().playSoundEffect("Click.ogg");
